refactor(feed): split oversize widgets
