+++ apps/mobile/lib/features/feed/presentation/feed_section_states.dart
@@ -1,0 +1,58 @@
+part of 'feed_section.dart';
+
+class _FeedFooter extends StatelessWidget {
+  const _FeedFooter();
+
+  @override
+  Widget build(BuildContext context) {
+    return const Padding(
+      padding: EdgeInsets.symmetric(vertical: 24),
+      child: Center(
+        child: Text(
+          'Đã hiển thị tất cả',
+          style: TextStyle(
+            color: AppColors.bw500,
+            fontSize: 14,
+            fontWeight: FontWeight.w600,
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _EmptyState extends StatelessWidget {
+  const _EmptyState();
+
+  @override
+  Widget build(BuildContext context) {
+    return const Padding(
+      padding: EdgeInsets.symmetric(vertical: 60, horizontal: 40),
+      child: Column(
+        children: [
+          Icon(Icons.photo_camera_outlined, color: AppColors.bw600, size: 48),
+          SizedBox(height: 16),
+          Text(
+            'Chưa có ảnh nào',
+            style: TextStyle(
+              color: AppColors.bw400,
+              fontSize: 18,
+              fontWeight: FontWeight.w800,
+              fontFamily: 'Nunito',
+            ),
+          ),
+          SizedBox(height: 8),
+          Text(
+            'Chụp ảnh đầu tiên và gửi cho bạn bè!',
+            style: TextStyle(
+              color: AppColors.bw600,
+              fontSize: 15,
+              fontWeight: FontWeight.w600,
+            ),
+            textAlign: TextAlign.center,
+          ),
+        ],
+      ),
+    );
+  }
+}
