+++ apps/mobile/lib/features/feed/presentation/home_screen.dart
@@ -23,6 +23,8 @@ import 'package:meep/shared/models/post.dart';
 import 'package:meep/shared/widgets/app_avatar.dart';
 import 'package:meep/shared/widgets/app_taskbar.dart';
 
+part 'home_screen_widgets.dart';
+
 class HomeScreen extends ConsumerStatefulWidget {
   const HomeScreen({
     super.key,
@@ -207,56 +209,9 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
     }
   }
 
-  /// Derive màu ring cho AppTaskbar từ Space đầu tiên trong post hiện tại.
-  /// Multi-Space post pick first (visual hint dùng 1 màu).
-  Color? get _ringColor {
-    if (_currentPage < 1 || _currentPage - 1 >= _posts.length) return null;
-    final post = _posts[_currentPage - 1];
-    if (post.spaceIds.isEmpty) return null;
-    final space = ref.read(spaceByIdProvider(post.spaceIds.first)).valueOrNull;
-    if (space == null) return null;
-    return hexToColor(space.colorHex);
-  }
-
   // TODO(Feed/KhoaLND): wire luồng chia sẻ — nút phải embedded là nút share,
   // chưa chốt đích đến.
   void _onShareTap() {}
-
-  /// Camera page (page 0) -> floating pill (hug-width, có active state).
-  /// Feed (page >=1) -> embedded full-width với nút grid + share ngoài pill.
-  Widget _buildTaskbar(int chatBadgeCount, Color? ringColor) {
-    if (_currentPage >= 1) {
-      return Padding(
-        padding: EdgeInsets.fromLTRB(
-          16,
-          12,
-          16,
-          MediaQuery.of(context).size.height * 0.04,
-        ),
-        child: AppTaskbar(
-          variant: TaskbarVariant.embedded,
-          activeTab: TaskbarTab.home,
-          chatBadgeCount: chatBadgeCount,
-          onTabSelected: _onTaskbarTab,
-          onGridTap: () => context.push('/grid-view'),
-          onUploadTap: _onShareTap,
-          ringColor: ringColor,
-        ),
-      );
-    }
-    return Padding(
-      padding: EdgeInsets.only(
-        bottom: MediaQuery.of(context).size.height * 0.045,
-      ),
-      child: Center(
-        child: AppTaskbar(
-          activeTab: TaskbarTab.home,
-          chatBadgeCount: chatBadgeCount,
-          onTabSelected: _onTaskbarTab,
-        ),
-      ),
-    );
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -337,299 +292,6 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
             _buildTaskbar(totalUnread, _ringColor),
           ],
         ),
-      ),
-    );
-  }
-
-  Widget _buildPageView(
-    List<Post>? posts, {
-    bool isLoading = false,
-    bool isError = false,
-  }) {
-    final postCount = (isLoading || isError || posts == null || posts.isEmpty)
-        ? 1
-        : posts.length;
-
-    return PageView.builder(
-      controller: _pageController,
-      scrollDirection: Axis.vertical,
-      onPageChanged: _onPageChanged,
-      itemCount: 1 + postCount,
-      itemBuilder: (context, index) {
-        if (index == 0) return CameraSection(onGoToFeed: _goToFeed);
-        if (isLoading) {
-          return const Center(
-            child: CircularProgressIndicator(color: AppColors.turquoise500),
-          );
-        }
-        if (isError) {
-          return FeedErrorView(
-            onRetry: () {
-              final sel = ref.read(feedFilterControllerProvider);
-              ref.invalidate(
-                feedControllerProvider(
-                  filter: _filterModeFor(sel),
-                  filterUid: sel.authorUid,
-                  filterSpaceId: sel.spaceId,
-                ),
-              );
-            },
-          );
-        }
-        if (posts == null || posts.isEmpty) {
-          return _EmptyFeedPage(
-            onCapture: () => _pageController.jumpToPage(0),
-          );
-        }
-        return _PostPage(post: posts[index - 1]);
-      },
-    );
-  }
-}
-
-class _HomeTopBar extends ConsumerWidget {
-  const _HomeTopBar({
-    required this.isFeedMode,
-    required this.selectedLabel,
-    required this.onFilterSelected,
-    required this.onSpaceFilterSelected,
-    this.spaceId,
-  });
-
-  /// true = viewing feed → show "Mọi người ▾" filter dropdown.
-  /// false = camera page → show "N bạn bè" pill that opens the FriendSheet.
-  final bool isFeedMode;
-
-  /// Label shown on the filter button ("Mọi người", "Bạn", or a friend name).
-  final String selectedLabel;
-
-  /// Called with (authorUid, label). authorUid null = "Mọi người".
-  final void Function(String? authorUid, String label) onFilterSelected;
-
-  /// Called khi user chọn Space từ dropdown — set local space filter +
-  /// sync currentSpaceProvider cho Camera context.
-  final void Function(Space space) onSpaceFilterSelected;
-
-  /// Non-null khi HomeScreen mở qua deeplink `/space/:spaceId`. Topbar
-  /// thay slot avatar bằng "..." icon mở [SpaceManagementSheet] (T10b).
-  /// Trong Space view, Settings không access trực tiếp từ topbar — user
-  /// vẫn vào được qua Profile tab.
-  final String? spaceId;
-
-  void _openFriendSheet(BuildContext context) {
-    // Modal sheet đồng bộ với các sheet khác (kéo xuống để đóng, tap barrier
-    // để đóng). FriendSheet render AppBottomSheet bên trong.
-    showModalBottomSheet<void>(
-      context: context,
-      isScrollControlled: true,
-      backgroundColor: Colors.transparent,
-      barrierColor: const Color(0x73000000),
-      builder: (_) => const FriendSheet(),
-    );
-  }
-
-  Future<void> _openFilterDropdown(
-    BuildContext context,
-    String currentUid,
-  ) async {
-    await showGeneralDialog<void>(
-      context: context,
-      barrierDismissible: true,
-      barrierLabel: 'Đóng bộ lọc',
-      barrierColor: const Color(0x73000000),
-      transitionDuration: const Duration(milliseconds: 150),
-      pageBuilder: (_, __, ___) => FeedFilterDropdown(
-        currentUid: currentUid,
-        selectedLabel: selectedLabel,
-        onFilterSelected: onFilterSelected,
-        onSpaceFilterSelected: onSpaceFilterSelected,
-      ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final currentUid = ref.watch(currentUidProvider).valueOrNull;
-    final friendCount = currentUid == null
-        ? 0
-        : ref.watch(
-            friendControllerProvider(currentUid)
-                .select((s) => s.friends.length),
-          );
-
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-      child: Row(
-        children: [
-          const SizedBox(width: 40),
-          const Spacer(),
-          if (isFeedMode)
-            _buildFilterButton(context, currentUid)
-          else
-            _buildFriendCountPill(context, friendCount),
-          const Spacer(),
-          if (spaceId != null)
-            Semantics(
-              button: true,
-              label: 'Quản lý Space',
-              child: GestureDetector(
-                onTap: () => SpaceManagementSheet.show(context, spaceId!),
-                child: Container(
-                  width: 40,
-                  height: 40,
-                  decoration: BoxDecoration(
-                    color: AppColors.bw700.withValues(alpha: 0.08),
-                    shape: BoxShape.circle,
-                  ),
-                  alignment: Alignment.center,
-                  child: const Icon(
-                    Icons.more_horiz,
-                    color: AppColors.bw100,
-                    size: 22,
-                  ),
-                ),
-              ),
-            )
-          else
-            const AppTopAvatar(),
-        ],
-      ),
-    );
-  }
-
-  /// Camera page: pill hiển thị số bạn bè thật.
-  /// Tap = mở FriendSheet (Figma). Space context được pick từ AudienceRow ở
-  /// CapturePreview thay vì long-press ở camera page.
-  Widget _buildFriendCountPill(BuildContext context, int friendCount) {
-    final label = friendCount == 0 ? 'Chưa có bạn bè' : '$friendCount bạn bè';
-    return GestureDetector(
-      onTap: () => _openFriendSheet(context),
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-        decoration: BoxDecoration(
-          color: AppColors.bw700.withValues(alpha: 0.08),
-          borderRadius: BorderRadius.circular(40),
-        ),
-        child: Row(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            SvgPicture.asset(
-              'assets/icons/ic_user_round.svg',
-              width: 18,
-              height: 18,
-              colorFilter: const ColorFilter.mode(
-                AppColors.bw100,
-                BlendMode.srcIn,
-              ),
-            ),
-            const SizedBox(width: 8),
-            Text(
-              label,
-              style: AppTextStyles.mdBold.copyWith(color: AppColors.bw100),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  /// Feed page: "Mọi người ▾" mở dropdown lọc theo người đăng.
-  Widget _buildFilterButton(BuildContext context, String? currentUid) {
-    return GestureDetector(
-      onTap: currentUid == null
-          ? null
-          : () => _openFilterDropdown(context, currentUid),
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-        decoration: BoxDecoration(
-          color: AppColors.bw700.withValues(alpha: 0.08),
-          borderRadius: BorderRadius.circular(40),
-        ),
-        child: Row(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Text(
-              selectedLabel,
-              style: AppTextStyles.mdBold.copyWith(color: AppColors.bw100),
-            ),
-            const SizedBox(width: 4),
-            const Icon(
-              Icons.keyboard_arrow_down,
-              color: AppColors.bw200,
-              size: 18,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-/// Full-screen page render từng post trong PageView. Background đổi theo
-/// `post.spaceId.colorHex` — paint ở LEVEL Scaffold (_scaffoldBgColor) thay
-/// vì ColoredBox riêng cho từng page, để header pill + footer taskbar
-/// (transparent bg) phủ cùng màu space xuyên qua.
-class _PostPage extends ConsumerWidget {
-  const _PostPage({required this.post});
-
-  final Post post;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final currentUid = ref.watch(currentUidProvider).valueOrNull;
-    return post.authorId == currentUid
-        ? OwnPostPage(post: post)
-        : FriendPostPage(post: post);
-  }
-}
-
-class _EmptyFeedPage extends StatelessWidget {
-  const _EmptyFeedPage({required this.onCapture});
-
-  /// Đưa user về trang camera (page 0) để chụp ảnh đầu tiên.
-  final VoidCallback onCapture;
-
-  @override
-  Widget build(BuildContext context) {
-    return Center(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          const Icon(
-            Icons.photo_camera_outlined,
-            color: AppColors.bw600,
-            size: 48,
-          ),
-          const SizedBox(height: 16),
-          const Text(
-            'Chưa có ảnh nào',
-            style: TextStyle(
-              color: AppColors.bw400,
-              fontSize: 18,
-              fontWeight: FontWeight.w800,
-              fontFamily: 'Nunito',
-            ),
-          ),
-          const SizedBox(height: 8),
-          const Text(
-            'Chụp ảnh đầu tiên và gửi cho bạn bè!',
-            style: TextStyle(
-              color: AppColors.bw600,
-              fontSize: 15,
-              fontWeight: FontWeight.w600,
-            ),
-            textAlign: TextAlign.center,
-          ),
-          const SizedBox(height: 20),
-          TextButton.icon(
-            onPressed: onCapture,
-            icon: const Icon(Icons.photo_camera, size: 18),
-            label: const Text('Chụp ảnh đầu tiên'),
-            style: TextButton.styleFrom(
-              foregroundColor: AppColors.turquoise500,
-            ),
-          ),
-        ],
       ),
     );
   }
